splitCode
