+++ src/utils/getSessionStorage.js
@@ -1,0 +1,9 @@
+export function getSessionStorage() {
+    const tokenFromLocalStorage = sessionStorage.getItem("spotifyToken");
+    if (tokenFromLocalStorage !== null) {
+      return tokenFromLocalStorage;
+    }
+    return false;
+  }
+  
+
